+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog
 - Auto inputting of room code from url's hash.
   - Auto joining of room if default username is set.
 - Countdown before quiz is started.
+- Button clear to the suggestion list.
 
 ### Changed
 - Link with room code in hash is put in clipboard rather than just the room